omap2+: clock: do not set hw_auto in clk_enable for ducati
The Ducati SS, being a processor sub-system, doesn't quite fit
the current hwmod model. Ducati has three hwmod devices including
two pseudo-hwmod devices for the reset control of the ducati
cores, and needs the clocks to be enabled while releasing the
resets.

The OMAP-specific clock enable code currently puts a clock-domain
into HW_AUTO mode automatically when a clock is enabled. The
remoteproc code has to perform a wakeup on the clock-domain before
enabling the remote processors because of this auto-setting to
make sure the reset releases can be effective.

A check has been added for Ducati to prevent the clock enable code
to enter HW_AUTO for Ducati, and the remoteproc code will manage
the HW_AUTO setting for this domain while sequencing the resets
properly.

Change-Id: I22195e29968e243dd149da4cf95c878f772dae67
Signed-off-by: Suman Anna <s-anna@ti.com>