Mitigate suspend related battery charge issues.
The coulomb counter doesn't seem to function correctly while the
device is in suspend; accumulator values are much too high, though
the sampler counter seems to be correct.

The original TI driver that this one is based on turns off the
current measurement while the device is in suspend. Adding that
logic to this driver restores the ability to estimate charge from
the voltage, since the outrageously large readings don't happen
anymore.

Refactored the gas gauge code a bit to make it somewhat more
legible.

Sign extension from the TI code was completely broken. Fixed it.

Change-Id: I500929773e278156224fea79be676c54c11e2685