Enable serial TTY lines on USB connector on kernel bootup
If an adb usb cable is not plugged into the usb port, 2 gpio lines
control what goes out the D+/D- lines of the connector.  The choices
are floating, console tty, usb or stereo audio.  To have Russ's
special board work by default, select console tty instead of floating.

If an adb usb cable is plugged in, it is always the USB signals
because of the 5V on the VBUS line.

Change-Id: I7b3133fc3c3c0618982f7108e9102a041114b9f3