notle: rework restart and poweroff paths.
The old restart/poweroff path hooked the notifier chain which should
not be used for the realization of low-level restart/poweroff because
the notifier chain is called before driver shutdown occurs.

Enabled twl6030-poweroff to take care of the poweroff path.

Implemented handler for arm_pm_restart to take care of the restart
logic, including recording the restart reason in SRAM and deciding
whether to perform cold or warm restart.

Change-Id: I6f9c0a7ba6ad6ebb913b46c3f333843e76164eee