Changes to support external gas gauge on EVT2.
- Split twl6030_notle_battery into twl6030_charger and twl6030_battery.
- Update twl-core to support old unified bci/notle driver or split drivers.
- Update board-notle to support EVT1 and EVT2 charger/battery configurations.

Due to the split between twl6030_battery and twl6030_charger, battery monitoring
while charging on EVT1 is even more approximate than before. This is due to the
battery monitor not being able to controll the charger to disable charging while
measuring vbat. The battery monitor should behave the same while discharging and
once the battery reaches full and the coulomb counter engages.

EVT2 is not adversely affected by this split as the external gas gauge completely
manages measuring the state of charge independently of the charger driver.

Change-Id: I84c0a559c42c742d5c1a77223647720a216b817a