Fix battery monitoring for EVT1.
- Add the gas gauge device only if board id is EVT2.
- Minor fix to legacy PMIC battery driver to report
  charging/discharging state based on average current sign.

It appears that the external gas gauge driver binds whether or not
the device is found on the bus. Having the external gas gauge driver
and the legacy PMIC battery driver both show up confuses the Android
layer, which seems to use the first battery driver it sees. If this
happens to be the external gas gauge driver on an EVT1 then Android
thinks there is no battery.

Change-Id: I696a17247465fa649361c80147ac1813256ffce6