Fix PER_DPLL M3 to 19.4
PER_DPLL M3 goes to ducati as a base clock to run the sensor.
It has to be 19.4 all the time for intershot timing to be 33ms in bursts
Before this change the M3 divisor would be adjusted and delays of 100ms would be introduced into burst capture
The 19.4 number is custom - normal clock is 19.2. We had to adjust ducati clock to shift Display clock so that it does not interfere with GPS signal

Change-Id: I7387e5f3857ba2d0749a6057bb575ba0e2f98107