Ignore interrupts when in boot mode. Make firmware flash more robust.
Joey and Baby Koala units share the IRQ between ALS and glasshub.
ALS is now generating interrupts which is causing the glasshub
IRQ handler to run. Glasshub doesn't support the IRQ status register
in boot mode, so don't schedule the IRQ handler if it is in boot
mode.

Bootloader version 4 adds checksums to the firmware flash message.
This makes the flash process more robust. The app processor can
now query if a page was flashed successfully. This change is
backwards compatible with version 1 & 3 bootloaders (V2 is borked).

Change-Id: I3dc1a3c46f20fa0aeb3b280482d9a75cef861fc6