Check VBUS state even when battery driver is not up yet.
Fix a timing related bug where the battery driver is not up yet
when this driver needs to evaluate the SoC and VBUS. Even if the
battery driver is not yet available, the VBUS state should be
checked.

In particular, charger mode in Android is sensitive to having the
correct state of VBUS early. When the charger daemon fails to see
USB online in the first 10 seconds the system is shutdown and does
not charge.

Change-Id: I1fa54355b857f1b5042f30d11c402aa751635a8b