Cache gpio key value during suspend
When suspended the time between pressing the camera button and
checking the gpio value may exceed the time the finger is still
pressing the button causing a camera input event not to be sent
to the input module.  Under these conditions the gpio keys will
generate a synthesized correct read and pass to the input module.

Change-Id: I4027a591446b3782d0fefe98b639f33d7d03a17a