Display changes for EVT2
This has a number of changes for display related to EVT2.  There are
new magic values from Adrian for led power settings and percentages
to set the whitepoint.  Enable a gamma table of 1.3 using the OMAP
gamma hardware.  Enable a CPR (Color Phase Rotation) matrix with
values determined by Marc to be used with the gamma of 1.3. Has
been testing with old 0x18 version of FPGA and Adrian's new 0x36
version.

Change-Id: I27aa8b099d0867f05e778a6b1e487b235933c58e