Fix prox sensor timestamp issue
We can see significant jitter in the prox data timestamps due to
latency in servicing the interrupt. This change always "fakes"
the timestamps by adding the sampling interval to the last
timestamp. The interval is adjusted once a second based on the
actual timestamps. This allows the system to adapt gradually to
reasonable amounts of drift. If the drift exceeds one sample,
there is a secondary correction that will gradually adjust the
fake timestamps to line up with the real-time data.

This change also deprecates MCU firmware older than V0.12. This
will prevent factory EVT2 units from working until the MCU firmware
is upgraded.

Change-Id: Ie548351b02454f9c3700fe885423df055dfda4a2