Fix flash when screen powers on
Turn gamma enable off when suspending.  Then on the resume, the
screen will not flash when LED is turned on.  The resume flash
was because the gamma table isn't loaded until after LED is turned
on.  Now, the next call after LED is on will load the table and
enable gamma.

Change-Id: I964e6b2830e0d5e642418a8f6543c37988862a33
NOTE: this requires FPGA image 0x38 or later to work