Add support for wink flag. Fix soft reboot issue.
This change plumbs a flag for each prox frame up to user space that
indicates if the low-power detector signaled a positive detection.
This will help with power so we can decide when to shut down high
power mode.

This also fixes an issue with soft reboot. A recent change to the
MCU firmware means it takes a bit more time to shut down the app
code which means a delay before the boot code will respond to
commands. Increase the delay to 50ms to allow time for the MCU
app code to tear down.

Change-Id: Iac0233b6526610440932346cf1b52bdb9e208541