Enable FIQ debugger.
With the FIQ debugger enabled, the console uart is now owned by the debugger.
By default, the uart console will be in FIQ debugger mode on boot. It can
be switched to the console mode with the "console" command at the debug prompt:

debug> console

In the console mode, a "break" can be sent via the terminal application to flip
the console uart back to the FIQ Debugger.

Please type "help" in the FIQ debugger for debug information.

Change-Id: I2b1241b816dcd30b5e0ae5bd381782cbf9a1e798