Don't register/unregister glasshub sysfs nodes
The glasshub driver previously registered and unregistered sysfs
nodes based on the device state. Unfortunately, this precludes the
use of "on init" in the init.rc file to chown the nodes to user
"system", since when they are recreated dynamically, they will
be owned by root. For security reasons, we don't want them make
them world read/write.

This change registered all the nodes at module probe, but disables
the device if down-rev firmware is detected. The app processor
can then re-enable the device and update the firmware.

Change-Id: I061e07b2c0f249890ff6959658b997aa1f80e326