Change the time to empty average property use a different calculation.
Battery time to empty based on one second average current is not very
useful. Change the property to read the time to empty based on average
standby current. This gives an upper bound on the TTE value and will
intuitively reflect heavy current draw, since power remaining is
factored into the calculation.

Change-Id: I12bab763041b5606edfd208c7128ebf7ce1ef794