Introduce USB MUX driver for the notle board.
First stab at centralizing the USB MUX control. This driver has
the following properties:

- Forces MUX to USB mode whenever VBUS is valid, regardless of
  userspace requested mode.
- Provides a path for the charger to also force MUX into USB mode
  before starting charger. This mitigates VBUS voltage droop
  problems.
- Provides a consistent sysfs entry for userspace to request the
  MUX mode while mainting overriding constraints.
- Provides framework to pave the way for automatically configuring
  the MUX based on USB ID -- to be implemented.

Userspace can request the MUX configuation by writing to the
following sysfs node:

	/sys/devices/platform/usb_mux.0/mux_mode

The sysfs node recognizes the following values:

	floating
	usb
	tty
	audio

The usersapce request will be honored whenever VBUS is not valid or
when the charger has not requested the mode to be forced to usb.
The requested mode is automatically resumed whenever an overriding
condition is not present.

Change-Id: I03d19a33ec73688c99495feb4fc9ca1a165b74a0