Fix I2C-4 wedge issue
The glasshub driver uses a threaded interrupt handler to do
the heavy lifting, since I2C transactions can't occur from
the top-half of the interrupt handler. The device driver can
be disabled at any time from user space (synced through a
mutex), which in turn resets the glasshub processor. In its
reset state, most I2C commands do not function, including
the IRQ status reqister query.

The top half of the driver checks this state and won't
schedule the bottom half if the device is reset. However,
there is a window where an interrupt could occur, the top-half
could run, schedule the bottom half, and then user space could
disable the driver and reset the chip. When the bottom half
runs, it tries to read the IRQ register, which hangs the I2C
bus.

This patch checks the driver disable bit before attempting any
I2C transactions from the bottom-half of the driver. When the
device is reset, and the IRQ line should be cleared, but the
bottom-half may have already been scheduled. We also print a
warning message, which should help confirm that this situation
does actually occur.

Once we retire Joey devices, as an additional check, we could
also read the status of the IRQ line and confirm that the
interrupt is no longer pending. This doesn't work on Joey
devices because they share an IRQ line with the ALS.

Change-Id: Ic0e1063823322319fda07064667fef8d6b91cb1d