Fix race between exposing battery to power supply core and updating reg cache.
The battery "detection" logic used to determine if VBAT is connected to a
battery or not depends on the values in the reg cache being valid. During
init it is possible for the power supply core to query the battery
properties before the first update of the reg cache. This results in a false
fake battery detection and the battery driver subquently always reports 96%.

Prefill the temperature regs before registering with power supply core to
avoid this race.

Change-Id: Ib00aaa88e859e9f6be0c86e1d62b73ebe3f7a2e4