Add support for factory cable
If factory cable (VBUS connected to ID) is inserted, then
reboot into fastboot mode.  Special androidbootmode of
factorycable will override the reboot (and a change in
Android will allow adb for user builds).  If cable is
unplugged in this mode, system will power off in
30 seconds.

For additional details, see bug 8620925

Change-Id: Idc7a4457b07b56fa87eb84c6d9445b11485b1581