Reboot instead of powering off if VBUS is detected.
The normal power off path is taken up until the last step to trigger
the PMIC to power off. There the VBUS status is checked and if present
the device is rebooted instead. The normal power off sequence ensures
that the reboot reason is set to "off" which will be seen by the
bootloader and result in booting into charger mode with an extra
parameter to disable autoboot.

Change-Id: I12f569e88c75fc16fff6cef0e6a7a0a1c015a0ee