Only send one synthesized key event per suspend/resume cycle
By the time the gpio is polled the key that generated the
initial interrupt could be released.  So synthesize an event
to generate the missed keypress.  However, this approach
bypasses any software button debounce checks so provide
an equivalent mechanism to only send a synthesized event
once per suspend resume cycle.

Change-Id: I06dfe9641091211b9011d1db91f56860cb83785f