Wait for L4CFG module clock to be operational before reg access.
This is a targeted fix for what seems to be a hardware race condition. The
USB PHY control register is part of the SYSCTRL CONTROL module and is in
the L4CFG clock domain. The module is setup for autoidle, so the clock gating
should be completely under hardware control, however it appears that the
module is not always operational by the time the USB PHY register is accessed.

This fix adds a check to wait for the module to be in operational state before
accessing the register. In testing this appears to mitigate the problem.

Worst case this fix will cause the enable phy routine to exit with an error due
to timing out instead of faulting due to accessing an unclocked module.

Change-Id: I559ce2392a3ef470affe1c5ff918a30c598e7791