Send poweroff event into userspace on SOC_INT + SYSDOWN flag.
The flags.SYSDOWN bit indicates that the voltage at the battery has
been below a programmable threshold for a programmable period. Use this
condition to send an event into userspace via switch_dev to signal that
Android should gracefully shutdown immediately.

The SYSDOWN bit is checked both on the normal polling operation and when
SOC_INT fires. SOC_INT is configured to be a wake source.

Change-Id: I41cb6712b64c898ec664b71e94e8f52b697f3120