New kernel option to keep USB mux in tty mode
There is a new kernel command line option that changes how
the USB mux works.  It is now possible to keep the mux
set to the tty so that people can get at the serial console.
The audio HAL was changed to force the mux to audio so that
the earbuds do not get the annoying pops and clicks from
kernel console messages.  The option "usb_mux.tty_priority=true"
on the command line will keep the mux on tty and ignore
requests to switch to audio.  Of course, this is for
debugging only and will not work with earbuds.  A change to
u-boot to use this new feature will be next.  For now you
can change the options to mkbootimg when you build a custom
boot.img.   b/9377498

Change-Id: Icf95a3aec8e6b80cd1e4749b585d777573669201