Remove unused shutdown code.
This is not useful because in order for us to get into factorycable
bootmode, we have to do 'fastboot oem reboot-factorycable' but that only
works when the device is unlocked. Since we have to lock the device
before it leaves usb for the last time, we have to lock from fastboot
and can no longer reboot-factorycable, only poweroff.

Change-Id: I42bce891884c61cd7fbc26daadce88d3c9754f4c