Enter Bluetooth low power (sniff-mode) on suspend
Previously, sniff-mode was requested only for connections that had
been idle for 500ms.  It was possible to suspend while the idle timer
was still pending.  In this case the BT chipset consumes power in
active mode while the host processor is suspended.  This is a possible
culprit for b/10950173.

After this change, we attempt to enter sniff mode on all connections
whenever the host processor is suspended.

Connections are then brought out of sniff mode on first touch.

Change-Id: I0aabc06f1f41e44e40938dcdb3a04cb81890f365