Auto-reconfigure FPGA on invalid revision
From the field we have seen rare case of systems where the display is
not working in an otherwise functioning system.  Hardware and software
probing suggests the FPGA is deconfigured (b/10948411).  As a
mitigation, detect this situation and reconfigure the FPGA.  Also print
a log message to help gather statistic on this failure mode.

NB: to force deconfigure of the FPGA on EVT2 and higher systems:

under adb shell:
omapconf write 0x48057190 0x00800000;
busybox usleep 1000;
omapconf write 0x48057194 0x00800000

(requires omapconf, present on eng builds)

Change-Id: Ib37ba186a0b3ccb2f6602bf749ba573e292ccd31