Horrible hack to fix I2C power management bug
If an IRQ occurs while we are trying to go into suspend, the I2C
driver can end up shutting down in the middle of an I2C transaction.
This hack bumps the usage_count of the I2C driver to prevent it
from going into suspend while we're trying to read data from the
MCU. Also adds a bit of debug output if a timeout ever occurs.

Change-Id: Ibb2bf27a88a0822e4b248746b10a4d47138b662b