Groundwork for brightmode lab
Perform a range expansion of app-land brightness requests so that 10-223
corresponds to current dim/normal brightness and 224-255 allows a ramp
between normal and extra-bright.  This change in itself has minimal
imapct on current device behavior but lays the ground work for an extra
bright mode and/or accessibility mode.

Works in conjunction with config.xml changes.

Change-Id: Ia99d752025ddbc92baa8b71905656f50a40754ab